backend/requirements.txt
Flask
tensorflow-cpu
numpy
Pillow
h5py